Added service for maintaining authors
Posts are now required to have an author and when the post is loaded the
author object is assinged to it.
Database collections now has the ability to be read from disk at a later
time so that, in this case, all authors are loaded before we start to load
any posts.